+++ lib/src/analyzers/lint_analyzer/rules/rules_list/use_setstate_synchronously/visitor.dart
@@ -44,6 +44,13 @@ class _AsyncSetStateVisitor extends RecursiveAstVisitor<void> {
   }
 
   @override
+  void visitAssertStatement(AssertStatement node) {
+    final newMounted = _extractMountedCheck(node.condition);
+    mounted = newMounted.or(mounted);
+    super.visitAssertStatement(node);
+  }
+
+  @override
   void visitMethodInvocation(MethodInvocation node) {
     if (!inAsync) {
       return node.visitChildren(this);
@@ -65,7 +72,6 @@ class _AsyncSetStateVisitor extends RecursiveAstVisitor<void> {
       return node.visitChildren(this);
     }
 
-    node.condition.visitChildren(this);
     final newMounted = _extractMountedCheck(node.condition);
     mounted = newMounted.or(mounted);
 
@@ -153,9 +159,8 @@ class _AsyncSetStateVisitor extends RecursiveAstVisitor<void> {
     final oldMounted = mounted;
     node.body.visitChildren(this);
     final afterBody = mounted;
-    // ignore: omit_local_variable_types
-    final MountedFact beforeCatch =
-        mounted == oldMounted ? oldMounted : false.asFact();
+    final beforeCatch =
+        mounted == oldMounted ? oldMounted : false.asFact<BinaryExpression>();
     for (final clause in node.catchClauses) {
       mounted = beforeCatch;
       clause.visitChildren(this);